Create Control panel  Code v1.0
Ardunio Code for two button control that uses PWM send power to lasers. In this beta I used a LED as the Laser. As the system is unable to give enought amps to power the laser, more likely I'll use a PNP Transistor to and the VIN to power the laser.